Confirmation (amendments) now saves payment details to keystore

+++ app/models/payments/PaymentDetails.scala
@@ -1,13 +1,15 @@
 package models.payments
 
+import models.confirmation.Currency
 import play.api.libs.json.Json
 
-case class PaymentDetails(reference: String, amount: Float)
+case class PaymentDetails(reference: String, amount: Double)
 
 object PaymentDetails {
 
   val cacheKey = "Payment_Confirmation_Details"
 
   implicit val formats = Json.format[PaymentDetails]
+
 
 }
